replaced <img> with gatsby static images removed LinkBox component solved lint warnings
