+++ src/components/CellLink/CellLink.js
@@ -1,0 +1,26 @@
+// Type
+//
+import {Cell} from '@enact/ui/Layout';
+import kind from '@enact/core/kind';
+import {Link} from 'gatsby';
+import PropTypes from 'prop-types';
+
+import css from './CellLink.module.less';
+
+const CellLink = kind({
+	name: 'CellLink',
+	propTypes: {
+		children: PropTypes.string,
+		to: PropTypes.string
+	},
+	styles: {
+		css,
+		className: 'cell'
+	},
+	render: ({to, children, ...rest}) => (
+		<Cell size="50%" component={Link} to={to} {...rest}>{children}</Cell>
+	)
+});
+
+export default CellLink;
+export {CellLink};
