+++ src/components/CellLink/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "CellLink.js"
+}
