--- src/components/LinkBox/package.json
@@ -1,3 +1,0 @@
-{
-	"main": "LinkBox.js"
-}
