+++ src/html.js
@@ -1,18 +1,18 @@
-import React from "react"
-import PropTypes from "prop-types"
+import React from 'react';
+import PropTypes from 'prop-types';
 
-export default function HTML(props) {
+export default function HTML (props) {
 	return (
-	    <html {...props.htmlAttributes}>
+		<html {...props.htmlAttributes}>
 			<head>
-				<script type="text/javascript" charSet="UTF-8" src="//cdn.cookie-script.com/s/3a846584c6b545a3d1ac4dcfc8ac15a2.js"></script>
+				<script type="text/javascript" charSet="UTF-8" src="//cdn.cookie-script.com/s/3a846584c6b545a3d1ac4dcfc8ac15a2.js" />
 				<meta charSet="utf-8" />
 				<meta httpEquiv="x-ua-compatible" content="ie=edge" />
 				<meta
 					name="viewport"
 					content="width=device-width, initial-scale=1, shrink-to-fit=no"
 				/>
-			    {props.headComponents}
+				{props.headComponents}
 			</head>
 			<body {...props.bodyAttributes}>
 				{props.preBodyComponents}
@@ -24,14 +24,14 @@ export default function HTML(props) {
 				{props.postBodyComponents}
 			</body>
 		</html>
-	)
+	);
 }
 
 HTML.propTypes = {
-	htmlAttributes: PropTypes.object,
-	headComponents: PropTypes.array,
-	bodyAttributes: PropTypes.object,
-	preBodyComponents: PropTypes.array,
 	body: PropTypes.string,
+	bodyAttributes: PropTypes.object,
+	headComponents: PropTypes.array,
+	htmlAttributes: PropTypes.object,
 	postBodyComponents: PropTypes.array,
-}
+	preBodyComponents: PropTypes.array
+};
