+++ src/pages/docs/developer-guide/index.js
@@ -1,18 +1,16 @@
 import {Helmet} from 'react-helmet';
 import {graphql} from 'gatsby';
+import {StaticImage as Image} from "gatsby-plugin-image";
 import PropTypes from 'prop-types';
 import {Component} from 'react';
 import {Row} from '@enact/ui/Layout';
 
-import {CellLink} from '../../../components/LinkBox';
+import CellLink from '../../../components/CellLink';
 import Page from '../../../components/DocsPage';
 import SiteTitle from '../../../components/SiteTitle';
 import SiteSection from '../../../components/SiteSection';
 
 import css from '../../../css/main.module.less';
-
-// images
-import guide from '../images/guide.svg';
 
 export const frontmatter = {
 	title: 'Developer Guide',
@@ -35,7 +33,7 @@ const Doc = class ReduxDocList extends Component {
 						<Helmet>
 							<meta name="description" content={frontmatter.description} />
 						</Helmet>
-						<h1 className={css.withCaption}><img alt="Location marked in a book" src={guide} />{frontmatter.title}</h1>
+						<h1 className={css.withCaption}><Image className={css.image} alt="Location marked in a book" loading="eager" placeholder="none" src="../images/guide.svg" />{frontmatter.title}</h1>
 						<div className={css.caption}>
 							<p>Details and resources on how to use Enact.</p>
 						</div>
