+++ src/pages/docs/index.module.less
@@ -1,8 +1,21 @@
 // index.module.less
 //
 
-// @import "../../css/colors.less";
+@import "../../css/colors.less";
 @import "../../css/variables.less";
+
+@keyframes wiggle {
+	0%, 5% {
+		transform: rotate(0deg);
+	}
+	1%, 3% {
+		transform: rotate(5deg);
+
+	}
+	2%, 4% {
+		transform: rotate(-5deg);
+	}
+}
 
 .gettingStarted {
 	font-size: 1.1em;
@@ -25,8 +38,10 @@
 		}
 
 		.image {
-			flex: 0 1 25%;
-			margin-bottom: 0;
+			img {
+				flex: 0 1 25%;
+				margin-bottom: 0;
+			}
 		}
 
 		.content {
@@ -34,7 +49,54 @@
 			margin-left: 5%;
 		}
 	}
+
+	.linkBox {
+		margin: 3em 0;
+
+		.imageContainer {
+			flex: 1 3;
+			margin-right: 2em;
+			text-align: center;
+
+			.image {
+				margin: 0 0 0.5em 0;
+			}
+
+			img {
+				width: 72px;
+				will-change: transform;
+				animation: 6s none infinite ease-in;
+				transform-origin: center bottom;
+			}
+		}
+
+		.contentCell {
+			flex: 3 1;
+		}
+
+		&:hover .image img {
+			animation-name: wiggle;
+		}
+	}
+
+	.content {
+		justify-content: space-between;
+	}
+
+	.cell {
+		min-width: 200px;
+		padding: 0.5em;
+		margin: 0.25em 0;
+
+		&:focus {
+			text-decoration: underline; // Both simple-shorthand (for old browsers), and longhand below for better browsers
+			text-decoration-color: @docs-link-focus-accent-color;
+			text-decoration-line: underline;
+			text-decoration-style: double;
+		}
+	}
 }
+
 @media only screen and (max-width: @docs-mobile-breakpoint) {
 	.gettingStarted {
 		.hero {
@@ -49,5 +111,19 @@
 				padding-left: 0;
 			}
 		}
+
+		.linkBox {
+			.content {
+				.cell {
+					padding-right: 0;
+					max-width: 100%;
+				}
+			}
+		}
 	}
 }
+
+
+
+
+
